stkWALLET: uint -> uint256

+++ contracts/stkWALLET.sol
@@ -2,10 +2,10 @@
 pragma solidity ^0.8.19;
 
 interface IXWallet {
-	function balanceOf(address owner) external view returns (uint);
-	function shareValue() external view returns (uint);
-	function transfer(address to, uint amount) external returns (bool);
-	function transferFrom(address from, address to, uint amount) external returns (bool);
+	function balanceOf(address owner) external view returns (uint256);
+	function shareValue() external view returns (uint256);
+	function transfer(address to, uint256 amount) external returns (bool);
+	function transferFrom(address from, address to, uint256 amount) external returns (bool);
 }
 
 contract stkWALLET {
@@ -19,29 +19,29 @@ contract stkWALLET {
 	IXWallet xWallet;
 
 	// Mutable variables
-	mapping(address => uint) public shares;
-	mapping(address => mapping(address => uint)) private allowed;
+	mapping(address => uint256) public shares;
+	mapping(address => mapping(address => uint256)) private allowed;
 
 	// ERC20 events
-	event Approval(address indexed owner, address indexed spender, uint amount);
-	event Transfer(address indexed from, address indexed to, uint amount);
+	event Approval(address indexed owner, address indexed spender, uint256 amount);
+	event Transfer(address indexed from, address indexed to, uint256 amount);
 	// Custom events
-	event ShareValueUpdate(uint shareValue);
+	event ShareValueUpdate(uint256 shareValue);
 
 	// ERC20 methods
 	// Note: any xWALLET sent to this contract will be burned as there's nothing that can be done with it. Expected behavior.
-	function totalSupply() external view returns (uint) {
+	function totalSupply() external view returns (uint256) {
 		return (xWallet.balanceOf(address(this)) * xWallet.shareValue()) / 1e18;
 	}
 
-	function balanceOf(address owner) external view returns (uint balance) {
+	function balanceOf(address owner) external view returns (uint256 balance) {
 		return (shares[owner] * xWallet.shareValue()) / 1e18;
 	}
 
-	function transfer(address to, uint amount) external returns (bool success) {
+	function transfer(address to, uint256 amount) external returns (bool success) {
 		require(to != address(this) && to != address(0), "BAD_ADDRESS");
-		uint shareValue = xWallet.shareValue();
-		uint sharesAmount = (amount * 1e18) / shareValue;
+		uint256 shareValue = xWallet.shareValue();
+		uint256 sharesAmount = (amount * 1e18) / shareValue;
 		shares[msg.sender] = shares[msg.sender] - sharesAmount;
 		shares[to] = shares[to] + sharesAmount;
 		emit Transfer(msg.sender, to, amount);
@@ -49,12 +49,12 @@ contract stkWALLET {
 		return true;
 	}
 
-	function transferFrom(address from, address to, uint amount) external returns (bool success) {
+	function transferFrom(address from, address to, uint256 amount) external returns (bool success) {
 		require(to != address(this) && to != address(0), "BAD_ADDRESS");
-		uint shareValue = xWallet.shareValue();
-		uint sharesAmount = (amount * 1e18) / shareValue;
+		uint256 shareValue = xWallet.shareValue();
+		uint256 sharesAmount = (amount * 1e18) / shareValue;
 		shares[from] = shares[from] - sharesAmount;
-		uint prevAllowance = allowed[from][msg.sender];
+		uint256 prevAllowance = allowed[from][msg.sender];
 		if (prevAllowance < type(uint256).max) allowed[from][msg.sender] = prevAllowance - amount;
 		shares[to] = shares[to] + sharesAmount;
 		emit Transfer(from, to, amount);
@@ -62,13 +62,13 @@ contract stkWALLET {
 		return true;
 	}
 
-	function approve(address spender, uint amount) external returns (bool success) {
+	function approve(address spender, uint256 amount) external returns (bool success) {
 		allowed[msg.sender][spender] = amount;
 		emit Approval(msg.sender, spender, amount);
 		return true;
 	}
 
-	function allowance(address owner, address spender) external view returns (uint remaining) {
+	function allowance(address owner, address spender) external view returns (uint256 remaining) {
 		return allowed[owner][spender];
 	}
 
@@ -83,14 +83,14 @@ contract stkWALLET {
 		wrap(xWallet.balanceOf(msg.sender));
 	}
 
-	function wrap(uint shareAmount) public {
+	function wrap(uint256 shareAmount) public {
 		shares[msg.sender] += shareAmount;
 		require(xWallet.transferFrom(msg.sender, address(this), shareAmount));
 		emit Transfer(address(0), msg.sender, (shareAmount * xWallet.shareValue()) / 1e18);
 	}
 
 	// this is used to trigger unstaking
-	function unwrap(uint shareAmount) external {
+	function unwrap(uint256 shareAmount) external {
 		shares[msg.sender] -= shareAmount;
 		require(xWallet.transfer(msg.sender, shareAmount));
 		emit Transfer(msg.sender, address(0), (shareAmount * xWallet.shareValue()) / 1e18);
